+++ CHANGELOG.md
@@ -8,6 +8,7 @@ First release.
 > Note that semantic versioning rules are not strictly followed during this phase.
 
 - feature/next-match: Filter: Add support for prev/next-match
+- feature/next-match: Filter: New mode 'mark' (like 'dim' but does not gray out)
 
 - v0.12.1: Fix flat source format for positional args.